+++ lib/cloudcannon-jekyll/generator.rb
@@ -2,6 +2,7 @@
 
 require "jekyll"
 require "fileutils"
+require_relative "reader"
 
 module CloudCannonJekyll
   # Generates JSON files containing build config and build output details
@@ -10,15 +11,15 @@ module CloudCannonJekyll
 
     def generate(site)
       @site = site
+      @reader = Reader.new(@site)
 
       collections_config = @site.config["collections"].dup || {}
-      drafts = read_drafts
 
       payload = @site.site_payload.merge({
         "gem_version" => CloudCannonJekyll::VERSION,
       })
 
-      add_blogging_config(collections_config, drafts)
+      drafts = add_blogging_config(collections_config)
       add_collection_paths(collections_config)
       add_data_config(collections_config)
 
@@ -34,6 +35,8 @@ module CloudCannonJekyll
     end
 
     def collections_dir
+      return "" if Jekyll::VERSION.start_with? "2."
+
       @site.config["collections_dir"] || ""
     end
 
@@ -41,42 +44,67 @@ module CloudCannonJekyll
       @site.config["data_dir"] || "_data"
     end
 
-    def read_data
-      if Jekyll::VERSION.start_with? "2."
-        CloudCannonJekyll::OldDataReader.new(@site).read("_data")
-      else
-        CloudCannonJekyll::DataReader.new(@site).read(data_dir)
-      end
-    end
+    def add_category_folder_config(collections_config, posts_config = {})
+      posts = @site.posts || @site.collections["posts"]
+      seen = {}
 
-    def read_drafts
-      if Jekyll::VERSION.start_with? "2."
-        @site.read_content("", "_drafts", Jekyll::Draft)
-      else
-        Jekyll::PostReader.new(@site).read_drafts(collections_dir)
+      posts.map do |post|
+        parts = post.relative_path.split("/_posts/")
+        path = parts.first
+
+        # Ignore unless it's an unseen category folder post
+        next if parts.length < 2 || path.empty? || seen[path]
+
+        # Could check this to ensure raw files exist since posts can be generated without files
+        # next if @reader.read_posts(parts[0]).empty?
+
+        seen[path] = true
+        folder = path.sub(%r!^\/+!, "")
+
+        collections_config["#{folder}/posts"] = posts_config.merge({
+          "_path" => "#{folder}/_posts",
+        })
+
+        # Adding the category draft config like this isn't ideal, since you could have drafts
+        #  without posts, but it's a decent trade off vs looking for _drafts folders
+        collections_config["#{folder}/drafts"] = posts_config.merge({
+          "_path" => "#{folder}/_drafts",
+        })
+
+        path
       end
     end
 
     # Add data to collections config if raw data files exist
-    def add_data_config(collections)
-      data_files = read_data
-      collections["data"] = { "_path" => data_dir } if data_files&.keys&.any?
+    def add_data_config(collections_config)
+      data_files = @reader.read_data(data_dir)
+      collections_config["data"] = { "_path" => data_dir } if data_files&.keys&.any?
     end
 
     # Add posts/drafts to collections config
-    def add_blogging_config(collections, drafts)
-      collections["posts"] = { "output" => true } if Jekyll::VERSION.start_with? "2."
+    def add_blogging_config(collections_config)
+      collections_config["posts"] = { "output" => true } if Jekyll::VERSION.start_with? "2."
+      drafts = @reader.read_drafts(collections_dir)
 
-      if collections.key?("posts")
-        collections["drafts"] = collections["posts"].dup
-      elsif drafts&.any?
-        collections["drafts"] = {}
+      if collections_config.key?("posts")
+        collections_config["drafts"] = collections_config["posts"].dup
+      elsif drafts.any?
+        collections_config["drafts"] = {}
       end
+
+      folders = add_category_folder_config(collections_config, collections_config["posts"])
+      folders.compact.each do |folder|
+        drafts += @reader.read_drafts(folder)
+      end
+
+      drafts
     end
 
     # Add _path to each collection config
-    def add_collection_paths(collections)
-      collections.each do |key, collection|
+    def add_collection_paths(collections_config)
+      collections_config.each do |key, collection|
+        next if collection.key?("_path")
+
         collection["_path"] = File.join(collections_dir, "_#{key}").sub(%r!^\/+!, "")
       end
     end