+++ HISTORY.md
@@ -2,6 +2,7 @@
 
 * Added drafts to collections in details
 * Added drafts and data to collections in config
+* Added category folder drafts and posts to collections in config
 * Added `_path` field to each collection definition in config
 * Removed some unused fields
 * Renamed `static` to `static-pages` in details and removed `robots.txt` and `sitemap.xml` exceptions